refactor: 프로젝트 구조 변경
- handleInput과 printResult 분리
- selectElements 분리
