+++ src/style.css
@@ -10,4 +10,5 @@
 #result-head span {
     animation-name: fontAnimation;
     animation-timing-function: ease-in-out;
+    animation-duration: 0s;
 }
